Rename `adminApi` package to `adminapi` for consistency across the codebase.
