+++ core/src/main/java/com/predic8/membrane/core/interceptor/adminapi/AdminApiInterceptor.java
@@ -11,7 +11,7 @@
    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
    See the License for the specific language governing permissions and
    limitations under the License. */
-package com.predic8.membrane.core.interceptor.adminApi;
+package com.predic8.membrane.core.interceptor.adminapi;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.ObjectMapper;
