+++ core/src/test/java/com/predic8/membrane/core/interceptor/adminapi/AdminApiInterceptorTest.java
@@ -11,7 +11,7 @@
    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
    See the License for the specific language governing permissions and
    limitations under the License. */
-package com.predic8.membrane.core.interceptor.adminApi;
+package com.predic8.membrane.core.interceptor.adminapi;
 
 import com.predic8.membrane.core.HttpRouter;
 import com.predic8.membrane.core.exchange.Exchange;
