+++ core/src/test/java/com/predic8/membrane/core/interceptor/adminapi/NotifyingByteArrayOutputStream.java
@@ -11,12 +11,9 @@
    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
    See the License for the specific language governing permissions and
    limitations under the License. */
-package com.predic8.membrane.core.interceptor.adminApi;
-
-import org.jetbrains.annotations.NotNull;
+package com.predic8.membrane.core.interceptor.adminapi;
 
 import java.io.ByteArrayOutputStream;
-import java.io.IOException;
 import java.util.concurrent.CountDownLatch;
 
 public class NotifyingByteArrayOutputStream extends ByteArrayOutputStream {
